Move Nitro API model to its own module
