+++ tools/container-converter/src/image_builder/parent.rs
@@ -8,7 +8,7 @@ use std::fs;
 use std::io::Write;
 use std::path::Path;
 
-use api_model::converter::NitroEnclavesConversionRequestOptions;
+use api_model::nitro::NitroEnclavesConversionRequestOptions;
 use docker_image_reference::Reference as DockerReference;
 use log::info;
 use tempfile::TempDir;
